web-auth-ldap/requirements.txt
flask
flask_wtf
email-validator
gunicorn
pycryptodome
pydantic
python-ldap
pycryptodome
